pyproject.toml
[tool.poetry]
name = "cq_yt_plugin"
version = "0.1.0"
description = ""
authors = ["Plaintextnerds <tim@plaintextnerds.com>"]
readme = "README.md"
packages = [{include = "plugin"}]

[tool.poetry.dependencies]
python = "^3.9,<3.13"
cloudquery-plugin-sdk = "^0.1.12"
pydantic = "^2.5"
pyarrow = "^14.0"
requests = "^2.31.0"
google-auth-oauthlib = "^1.2.0"
google-api-python-client = "^2.112.0"
google-auth-httplib2 = "^0.2.0"
python-dotenv = "^1.0.0"
imagehash = "^4.3.1"
pony = "^0.7.17"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"

[tool.poetry.scripts]
main = "main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

